requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.14.2
black==24.8.0
bleach==6.2.0
build==1.3.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.7
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
docopt==0.6.2
executing==2.2.1
fastapi==0.112.2
fastjsonschema==2.21.2
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
identify==2.6.14
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nodeenv==1.9.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pip==25.2
pip-tools==7.5.0
pipreqs==0.5.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.8.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.2.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruff==0.6.9
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.38.6
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.30.5
virtualenv==20.34.0
wcwidth==0.2.14
webencodings==0.5.1
wheel==0.45.1
yarg==0.1.9
